feat: filter on orgunits

See merge request rammearkitektur/oS2mo-os2sync-export!128

+++ os2sync_export/main.py
@@ -2,16 +2,21 @@
 #
 # SPDX-License-Identifier: MPL-2.0
 import logging
+from typing import Annotated
 from typing import Dict
 from uuid import UUID
 
 import sentry_sdk
 from fastapi import APIRouter
 from fastapi import BackgroundTasks
+from fastapi import Depends
 from fastapi import FastAPI
+from fastramqpi.depends import from_user_context
+from fastramqpi.depends import LegacyGraphQLSession
 from fastramqpi.main import FastRAMQPI  # type: ignore
+from gql import gql
 from gql.client import AsyncClientSession
-from ramqp.depends import Context
+from more_itertools import one
 from ramqp.depends import RateLimit
 from ramqp.mo import MORouter
 from ramqp.mo import PayloadUUID
@@ -33,13 +38,8 @@ logger = logging.getLogger(__name__)
 fastapi_router = APIRouter()
 amqp_router = MORouter()
 
-
-def unpack_context(context) -> tuple[Settings, AsyncClientSession, OS2SyncClient]:
-    """Returns the relevant objects from the context dictionary"""
-    settings: Settings = context["user_context"]["settings"]
-    graphql_session: AsyncClientSession = context["graphql_session"]
-    os2sync_client: OS2SyncClient = context["user_context"]["os2sync_client"]
-    return settings, graphql_session, os2sync_client
+Settings_ = Annotated[Settings, Depends(from_user_context("settings"))]
+OS2SyncClient_ = Annotated[OS2SyncClient, Depends(from_user_context("os2sync_client"))]
 
 
 @fastapi_router.get("/")
@@ -49,10 +49,11 @@ async def index() -> Dict[str, str]:
 
 @fastapi_router.post("/trigger", status_code=202)
 async def trigger_all(
-    context: Context, background_tasks: BackgroundTasks
+    background_tasks: BackgroundTasks,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
 ) -> Dict[str, str]:
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
-
     background_tasks.add_task(
         main,
         settings=settings,
@@ -64,10 +65,12 @@ async def trigger_all(
 
 @amqp_router.register("person")
 async def amqp_trigger_employee(
-    context: Context, uuid: PayloadUUID, _: RateLimit
+    uuid: PayloadUUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
+    _: RateLimit,
 ) -> None:
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
-
     try:
         sts_users = await get_sts_user(
             str(uuid),
@@ -83,15 +86,24 @@ async def amqp_trigger_employee(
 
 
 @amqp_router.register("org_unit")
-async def amqp_trigger_org_unit(context: Context, uuid: PayloadUUID, _: RateLimit):
-    settings, _, os2sync_client = unpack_context(context=context)
-
+async def amqp_trigger_org_unit(
+    uuid: PayloadUUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
+    _: RateLimit,
+) -> None:
     sts_org_unit = None
-    try:
-        sts_org_unit = get_sts_orgunit(str(uuid), settings=settings)
-    except ValueError:
-        logger.info(f"Event registered but no org_unit was found with {uuid=}")
-        os2sync_client.delete_orgunit(uuid)
+    if await is_below_top_unit(
+        graphql_session,
+        uuid,
+        settings.os2sync_top_unit_uuid,
+    ):
+        try:
+            sts_org_unit = get_sts_orgunit(uuid, settings=settings)
+        except ValueError:
+            logger.info(f"Event registered but no org_unit was found with {uuid=}")
+            os2sync_client.delete_orgunit(uuid)
     if sts_org_unit is None:
         os2sync_client.delete_orgunit(uuid)
         return
@@ -101,8 +113,13 @@ async def amqp_trigger_org_unit(context: Context, uuid: PayloadUUID, _: RateLimi
 
 
 @amqp_router.register("address")
-async def amqp_trigger_address(context: Context, uuid: PayloadUUID, _: RateLimit):
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
+async def amqp_trigger_address(
+    uuid: PayloadUUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
+    _: RateLimit,
+) -> None:
     try:
         ou_uuid, e_uuid = await get_address_org_unit_and_employee_uuids(
             graphql_session, uuid
@@ -111,7 +128,11 @@ async def amqp_trigger_address(context: Context, uuid: PayloadUUID, _: RateLimit
         logger.debug(f"No address found {uuid=}")
         return
 
-    if ou_uuid:
+    if ou_uuid and is_below_top_unit(
+        graphql_session,
+        uuid,
+        settings.os2sync_top_unit_uuid,
+    ):
         try:
             sts_org_unit = get_sts_orgunit(ou_uuid, settings)
         except ValueError:
@@ -133,10 +154,47 @@ async def amqp_trigger_address(context: Context, uuid: PayloadUUID, _: RateLimit
     )
 
 
-@amqp_router.register("ituser")
-async def amqp_trigger_it_user(context: Context, uuid: PayloadUUID, _: RateLimit):
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
+async def is_below_top_unit(
+    gql_session: AsyncClientSession,
+    unit_uuid: UUID,
+    top_unit_uuid: UUID,
+) -> bool:
+    """Checks whether an organisation unit is below the top unit.
 
+    This check is necessary because fk-org only supports one top level unit.
+    """
+    query = """
+    query QueryAncestors($uuids: [UUID!]) {
+        org_units(uuids: $uuids) {
+            current {
+            ancestors {
+                uuid
+            }
+            }
+        }
+    }
+     """
+    res = await gql_session.execute(
+        gql(query), variable_values={"uuids": str(unit_uuid)}
+    )
+    if not res["org_units"]:
+        logger.warn("No unit found")
+        return False
+    ancestors = one(res["org_units"])["current"]["ancestors"]
+    if ancestors is None:
+        return False
+    else:
+        return top_unit_uuid in {UUID(a["uuid"]) for a in ancestors}
+
+
+@amqp_router.register("ituser")
+async def amqp_trigger_it_user(
+    uuid: PayloadUUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
+    _: RateLimit,
+) -> None:
     try:
         ou_uuid, e_uuid = await get_ituser_org_unit_and_employee_uuids(
             graphql_session, uuid
@@ -145,7 +203,9 @@ async def amqp_trigger_it_user(context: Context, uuid: PayloadUUID, _: RateLimit
         logger.debug(f"Event registered but no it-user found with {uuid=}")
         return
 
-    if ou_uuid:
+    if ou_uuid and await is_below_top_unit(
+        graphql_session, ou_uuid, settings.os2sync_top_unit_uuid
+    ):
         try:
             sts_org_unit = get_sts_orgunit(ou_uuid, settings)
         except ValueError:
@@ -180,16 +240,24 @@ async def amqp_trigger_it_user(context: Context, uuid: PayloadUUID, _: RateLimit
 
 
 @amqp_router.register("manager")
-async def amqp_trigger_manager(context: Context, uuid: PayloadUUID, _: RateLimit):
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
-
+async def amqp_trigger_manager(
+    uuid: PayloadUUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
+    _: RateLimit,
+) -> None:
     try:
         ou_uuid = await get_manager_org_unit_uuid(graphql_session, uuid)
     except ValueError:
         logger.debug(f"Event registered but no manager found with {uuid=}")
         return
 
-    if ou_uuid:
+    if ou_uuid and await is_below_top_unit(
+        graphql_session,
+        ou_uuid,
+        settings.os2sync_top_unit_uuid,
+    ):
         try:
             sts_org_unit = get_sts_orgunit(ou_uuid, settings)
         except ValueError:
@@ -203,9 +271,13 @@ async def amqp_trigger_manager(context: Context, uuid: PayloadUUID, _: RateLimit
 
 
 @amqp_router.register("engagement")
-async def amqp_trigger_engagement(context: Context, uuid: PayloadUUID, _: RateLimit):
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
-
+async def amqp_trigger_engagement(
+    uuid: PayloadUUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
+    _: RateLimit,
+) -> None:
     try:
         e_uuid = await get_engagement_employee_uuid(graphql_session, uuid)
     except ValueError:
@@ -223,11 +295,19 @@ async def amqp_trigger_engagement(context: Context, uuid: PayloadUUID, _: RateLi
 
 
 @amqp_router.register("kle")
-async def amqp_trigger_kle(context: Context, uuid: PayloadUUID, _: RateLimit):
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
-
+async def amqp_trigger_kle(
+    uuid: PayloadUUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
+    _: RateLimit,
+) -> None:
     ou_uuid = await get_kle_org_unit_uuid(graphql_session, uuid)
-    if ou_uuid:
+    if ou_uuid and await is_below_top_unit(
+        graphql_session,
+        ou_uuid,
+        settings.os2sync_top_unit_uuid,
+    ):
         try:
             sts_org_unit = get_sts_orgunit(ou_uuid, settings)
         except ValueError:
@@ -242,11 +322,11 @@ async def amqp_trigger_kle(context: Context, uuid: PayloadUUID, _: RateLimit):
 
 @fastapi_router.post("/trigger/user/{uuid}")
 async def trigger_user(
-    context: Context,
     uuid: UUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
 ) -> str:
-    settings, graphql_session, os2sync_client = unpack_context(context=context)
-
     sts_users = await get_sts_user(
         str(uuid),
         gql_session=graphql_session,
@@ -260,12 +340,13 @@ async def trigger_user(
 
 @fastapi_router.post("/trigger/orgunit/{uuid}", status_code=200)
 async def trigger_orgunit(
-    context: Context,
     uuid: UUID,
+    settings: Settings_,
+    graphql_session: LegacyGraphQLSession,
+    os2sync_client: OS2SyncClient_,
 ) -> str:
-    settings, _, os2sync_client = unpack_context(context=context)
     try:
-        sts_org_unit = get_sts_orgunit(str(uuid), settings=settings)
+        sts_org_unit = get_sts_orgunit(uuid, settings=settings)
     except ValueError:
         logger.info("Org_unit not found")
         return "Org_unit not found"
